+++ dubbo-samples-attachment/pom.xml
@@ -29,5 +29,10 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>dubbo-samples-attachment</artifactId>
-
+    <dependencies>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context</artifactId>
+        </dependency>
+    </dependencies>
 </project>
