+++ dubbo-samples-attachment/src/main/java/org/apache/dubbo/samples/attachment/AttachmentConsumer.java
@@ -17,18 +17,17 @@
  *
  */
 
-package com.alibaba.dubbo.samples.attachment;
-
-import com.alibaba.dubbo.samples.attachment.api.AttachmentService;
+package org.apache.dubbo.samples.attachment;
 
 import org.apache.dubbo.rpc.RpcContext;
+import org.apache.dubbo.samples.attachment.api.AttachmentService;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 
 public class AttachmentConsumer {
 
     public static void main(String[] args) {
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/attachment-consumer.xml"});
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"META-INF/spring/attachment-consumer.xml"});
         context.start();
         AttachmentService attachmentService = (AttachmentService) context.getBean("demoService"); // get remote service proxy
 
@@ -39,6 +38,5 @@ public class AttachmentConsumer {
 
         hello = attachmentService.sayHello("world"); //attachment only affective once
         System.out.println(hello); // get result
-
     }
 }
