+++ dubbo-samples-attachment/src/main/resources/META-INF/spring/attachment-consumer.xml
@@ -33,7 +33,7 @@
 
     <!-- generate proxy for the remote service, then demoService can be used in the same way as the
     local regular interface -->
-    <dubbo:reference id="demoService" check="false" interface="com.alibaba.dubbo.samples.attachment.api.AttachmentService"/>
+    <dubbo:reference id="demoService" check="false" interface="org.apache.dubbo.samples.attachment.api.AttachmentService"/>
 
 </beans>
