+++ dubbo-samples-attachment/src/main/resources/META-INF/spring/attachment-provider.xml
@@ -34,9 +34,9 @@
     <dubbo:protocol name="dubbo" port="20890"/>
 
     <!-- service implementation, as same as regular local bean -->
-    <bean id="demoService" class="com.alibaba.dubbo.samples.attachment.impl.AttachmentImpl"/>
+    <bean id="demoService" class="org.apache.dubbo.samples.attachment.impl.AttachmentImpl"/>
 
     <!-- declare the service interface to be exported -->
-    <dubbo:service interface="com.alibaba.dubbo.samples.attachment.api.AttachmentService" ref="demoService"/>
+    <dubbo:service interface="org.apache.dubbo.samples.attachment.api.AttachmentService" ref="demoService"/>
 
 </beans>
